requirements.txt
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.35.0
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
toml==0.10.2
pydub
asyncio
aiofiles
flask==3.0.3
flask[async]
mysql.connector
flask_cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
bcrypt-4.2.0